refactor: Update private messages page and templates
This commit adds a new file, privateMessages.js, in the future this will handle the functionality of the private messages page. It also improves the design of the send private message page by adding the current user avatar and the receiver avatar. Additionally, it adds an option to attach an image to the message. This enhances the user experience and makes the private messages feature more robust.

Finally, the current user private messages page has been redesigned adding more information from each message and added an attached image section if the message contains one

In the future, users will be able to delete and reply to the messages